Update subscription descriptions and URLs
This commit modifies the text of the subscription options in the localization files for Belarusian, English, and Russian languages by removing specific pricing details for the monthly and gift subscriptions. Additionally, the URLs for the monthly purchase and gift subscription links on the EvoCoders page have been updated. The monthly subscription link now redirects to a new URL, and the gift subscription link directs to a new Telegram app link. Removed an unused annual subscription link from the EvoCoders page. These changes aim to streamline the purchasing descriptions and adjust the URLs for improved navigation.
⭐⭐⭐⭐⭐⭐⭐⭐⭐⭐